pagesの重複を削除
