+++ frontend/app/nextjs-blog/pages/index.js
@@ -1,8 +1,94 @@
+import TemplateTop from "../components/templates/top";
 
-import Top from '../components/pages/top'
 
-export default function Home() {
-		return (
-			<Top />
-		);
-	};
+
+const dummyPosts = [{
+		date: "Jun 1, 2020",
+		category: "Lravel",
+		title: "Build Your New Idea with Laravel Freamwork.",
+		text: "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempora expedita dicta totam aspernatur doloremque. Excepturi iste iusto eosenim reprehenderit nisi, accusamus delectus nihil quis facere in modi ratione libero!",
+		avatar: "https://images.unsplash.com/photo-1492562080023-ab3db95bfbce?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=731&q=80",
+		author_name: "Alex John",
+	},
+	{
+		date: "Feb 14, 2019",
+		category: "Design",
+		title: "Accessibility tools for designers and developers",
+		text: "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempora expedita dicta totam aspernatur doloremque. Excepturi iste iusto eosenim reprehenderit nisi, accusamus delectus nihil quis facere in modi ratione libero!",
+		avatar: "https://images.unsplash.com/photo-1464863979621-258859e62245?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=333&q=80",
+		author_name: "Jane Doe",
+	},
+	{
+		date: "Jun 1, 2020",
+		category: "Lravel",
+		title: "Build Your New Idea with Laravel Freamwork.",
+		text: "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempora expedita dicta totam aspernatur doloremque. Excepturi iste iusto eosenim reprehenderit nisi, accusamus delectus nihil quis facere in modi ratione libero!",
+		avatar: "https://images.unsplash.com/photo-1492562080023-ab3db95bfbce?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=731&q=80",
+		author_name: "Alex John",
+	},
+	{
+		date: "Feb 14, 2019",
+		category: "Design",
+		title: "Accessibility tools for designers and developers",
+		text: "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempora expedita dicta totam aspernatur doloremque. Excepturi iste iusto eosenim reprehenderit nisi, accusamus delectus nihil quis facere in modi ratione libero!",
+		avatar: "https://images.unsplash.com/photo-1464863979621-258859e62245?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=333&q=80",
+		author_name: "Jane Doe",
+	},
+];
+
+
+const dummyAuthors = [{
+		avatar: "https://images.unsplash.com/photo-1492562080023-ab3db95bfbce?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=731&q=80",
+		name: "Alex John",
+		post_count: 23,
+	},
+	{
+		avatar: "https://images.unsplash.com/photo-1464863979621-258859e62245?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=333&q=80",
+		name: "Jane Doe",
+		post_count: 52,
+	},
+];
+
+const dummyCategories = [{
+		name: "AWS"
+	},
+	{
+		name: "Laravel"
+	},
+	{
+		name: "Vue"
+	},
+	{
+		name: "Design"
+	},
+	{
+		name: "Django"
+	},
+	{
+		name: "PHP"
+	},
+];
+
+const dummyRecentPosts = [{
+	date: "Jun 1, 2020",
+	category: "Lravel",
+	title: "Build Your New Idea with Laravel Freamwork.",
+	avatar: "https://images.unsplash.com/photo-1492562080023-ab3db95bfbce?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=731&q=80",
+	author_name: "Alex John",
+}, ]
+
+export default function Top() {
+	return <TemplateTop posts = {
+		dummyPosts
+	}
+	authors = {
+		dummyAuthors
+	}
+	categories = {
+		dummyCategories
+	}
+	recentPosts = {
+		dummyRecentPosts
+	}
+	/>;
+}
